Core/Script: Updated several includes
	* Updated the copyright
	* Changed the directories of some custom scripts to manage them better
