+++ src/server/scripts/Custom/NPC's/npc_top_5_matadores.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 DeathCore <http://www.noffearrdeathproject.org/>
+ * Copyright (C) 2016-2017 DeathCore <http://www.noffearrdeathproject.org/>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -15,7 +15,7 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "../../../scripts/ScriptPCH.h"
+#include "ScriptPCH.h"
 
 /*Class E_Menu*/
 #define E_Warrior "Ver os top 5 Warriors com mais mortes"
