+++ src/server/scripts/Custom/NPC's/npc_weapon_transmog.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 DeathCore <http://www.noffearrdeathproject.org/>
+ * Copyright (C) 2016-2017 DeathCore <http://www.noffearrdeathproject.org/>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -24,7 +24,7 @@
  * VALUES('700010','0','0','0','0','0','736','0','0','0','Weapons Transmog NPC','BattleArenas','','0','80','80','1','35','35','1','1','1.14286','2','0','5000','5000','0','35000','10','1','0','1','512','0','0','0','0','0','0','10000','10000','68','0','0','0','0','0','0','0','0','0','0','0','0','0','0','0','0','0','0','0','0','0','0','0','','0','3','20','20','20','0','0','0','0','0','0','0','0','255','0','0','0','weptransmog','0');
 */
 
-#include "../../../scripts/ScriptPCH.h"
+#include "ScriptPCH.h"
 
 class weptransmog : public CreatureScript
 {
