+++ src/server/scripts/Custom/anti_divulgador.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 DeathCore <http://www.noffearrdeathproject.org/>
+ * Copyright (C) 2016-2017 DeathCore <http://www.noffearrdeathproject.org/>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -15,8 +15,8 @@
  * with this program. If not, see <http://www.gnu.org/licenses/>.
  */
  
-#include "../../../scripts/ScriptPCH.h"
-#include "../../../game/Chat/Channels/Channel.h"
+#include "ScriptPCH.h"
+#include "Channel.h"
  
 class System_Censure : public PlayerScript
 {
