+++ src/server/scripts/Custom/dobro_xp_evento.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 DeathCore <http://www.noffearrdeathproject.org/>
+ * Copyright (C) 2016-2017 DeathCore <http://www.noffearrdeathproject.org/>
  *
  * This program is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License as published by the
@@ -18,7 +18,6 @@
 #include "Player.h"
 #include "Chat.h"
 #include "World.h"
-#include "boost/date_time.hpp"
 #include "Config.h"
 
 class XpWeekend : public PlayerScript
